Extract old js/time constructs to separate file
Also:
* Rename new Day to Date
